fix file dir
