+++ Quackademics/src/components/Notes/Editor.jsx
@@ -1,0 +1,26 @@
+import React from "react";
+import {
+  MDXEditor,
+  headingsPlugin,
+  listsPlugin,
+  quotePlugin,
+  thematicBreakPlugin,
+  markdownShortcutPlugin,
+} from "@mdxeditor/editor";
+
+const Editor = () => {
+  return (
+    <MDXEditor
+      markdown={"# Hello World"}
+      plugin={[
+        headingsPlugin(),
+        listsPlugin(),
+        quotePlugin(),
+        thematicBreakPlugin(),
+        markdownShortcutPlugin(),
+      ]}
+    />
+  );
+};
+
+export default Editor;
