--- Quackademics/src/components/Notes/MDEditor/Editor.jsx
@@ -1,32 +1,0 @@
-import React from "react";
-import {
-  MDXEditor,
-  headingsPlugin,
-  listsPlugin,
-  quotePlugin,
-  thematicBreakPlugin,
-  markdownShortcutPlugin,
-} from "@mdxeditor/editor";
-import "@mdxeditor/editor/style.css";
-
-const Editor = () => {
-  return (
-    <h1>
-      Notes :3
-      <div>
-        MDEditor
-        <MDXEditor
-          plugin={[
-            headingsPlugin(),
-            listsPlugin(),
-            quotePlugin(),
-            thematicBreakPlugin(),
-            markdownShortcutPlugin(),
-          ]}
-        />
-      </div>
-    </h1>
-  );
-};
-
-export default Editor;
